backend/agent/requirements.txt
langchain
langchain-core
langgraph
langgraph-api

poetry
chainlit
matplotlib
numpy
python-dotenv
rich
typing-extensions
docker
ezdxf
boto3

langchain-google-genai
langchain_openai
langchain_anthropic

qdrant-client
sentence-transformers

